shifting of classes to new packages and additional splitting of firebase and local storage methods

+++ app/src/main/java/com/patrickgrady/todayslists/Managers/TasksManager.java
@@ -2,36 +2,55 @@ package com.patrickgrady.todayslists.Managers;
 
 import android.content.Context;
 import android.content.SharedPreferences;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 
-import com.patrickgrady.todayslists.Managers.Objects.ListOfTasks;
+import com.google.firebase.database.ChildEventListener;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.patrickgrady.todayslists.Objects.DailyTasks;
+import com.patrickgrady.todayslists.Objects.ListOfTasks;
+import com.patrickgrady.todayslists.Objects.firebase.ListOfTasksFire;
+import com.patrickgrady.todayslists.Objects.local.ListOfTasksLocal;
 
 import java.io.File;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.HashMap;
 import java.util.Iterator;
-import java.util.List;
 import java.util.Random;
 import java.util.TreeMap;
 import java.util.UUID;
 
 public class TasksManager extends TreeMap<String, ListOfTasks> implements UpdateManager.TimeSensitive {
 
+    // static ref
     private static TasksManager instance;
-    Context context;
-    ArrayList<String> order;
-    SharedPreferences pref;
+
+    // shared member variables
+    private DailyTasks dailyTasks;
+    private ArrayList<String> order;
+    private boolean localMode;
+
+    // local storage member variables
+    private SharedPreferences pref;
+    private File directory;
+
+    // firebase member variables
+    private DatabaseReference dataRef;
 
     //#region constructor/set-up
-    private TasksManager(Context c) {
-        context = c;
+    private TasksManager(boolean lm) {
+        localMode = lm;
         order = new ArrayList<>();
-        pref = context.getSharedPreferences("tasks",Context.MODE_PRIVATE);
     }
 
-    public static TasksManager getInstance(Context c) {
+    public static TasksManager getInstance(boolean lm, Context ...c) {
         if (instance == null) {
-            instance = new TasksManager(c);
-            instance.load();
+            instance = new TasksManager(lm);
+            instance.load(c);
         }
 
         return instance;
@@ -50,46 +69,45 @@ public class TasksManager extends TreeMap<String, ListOfTasks> implements Update
     @Override
     public void clear() {
         super.clear();
-        pref.edit().clear().commit();
+        dailyTasks.clear();
         order.clear();
-        for(String filename : fileNames()) {
-            new File(getDirectory(), filename).delete();
+
+        if(localMode) {
+            for (String filename : fileNames()) {
+                new File(getDirectory(), filename).delete();
+            }
         }
-        updateTodaysTasks();
     }
 
     public void addList() {
-        List<String> tasks = new ArrayList<>();
         String key = UUID.randomUUID().toString();
 
-        ListOfTasks t = new ListOfTasks(getDirectory(), key);
-        t.setAll(tasks);
+        ListOfTasks t = new ListOfTasksFire(key);
         this.put(key, t);
         order.add(key);
     }
 
     public void remove(String key) {
-        if(Arrays.asList(fileNames()).contains(key)) {
+        if(localMode && Arrays.asList(fileNames()).contains(key)) {
             new File(getDirectory(), key).delete();
         }
         super.remove(key);
-        pref.edit().remove(key).commit();
+        dailyTasks.remove(key);
         order.remove(key);
-        updateTodaysTasks();
     }
 
     private void remove(String key, Iterator itr) {
-        if(Arrays.asList(fileNames()).contains(key)) {
+        if(localMode && Arrays.asList(fileNames()).contains(key)) {
             new File(getDirectory(), key).delete();
         }
         super.remove(key);
-        pref.edit().remove(key).commit();
+        dailyTasks.remove(key);
         itr.remove();
     }
 
     // returns a list of today's tasks
     public ArrayList<String> getTasks() {
-        updateTodaysTasks();
+        refresh();
         return order;
     }
 
@@ -105,50 +123,37 @@ public class TasksManager extends TreeMap<String, ListOfTasks> implements Update
 
     //#region update daily tasks
     @Override
-    public void update() {
-        pref.edit().clear().commit();
-        updateTodaysTasks();
+    public void update(long updateTimeMillis) {
+        updateTodaysTasks(updateTimeMillis);
     }
 
     @Override
     public void refresh() {
-        updateTodaysTasks();
+        updateTodaysTasks(null);
     }
 
     // private helper/convenience methods
 
     // update today's chosen tasks list
-    private void updateTodaysTasks() {
+    private void updateTodaysTasks(Long updateTimeMillis) {
         // clear out empty task lists
         for(Iterator itr = order.iterator(); itr.hasNext();) {
             String key = (String) itr.next();
 
-            String currentTask = pref.getString(key, "");
+            String currentTask = dailyTasks.getTask(key);
             ListOfTasks tasks = this.get(key);
             if(tasks != null) {
                 if(tasks.isEmpty()) {
                     this.remove(key, itr);
                 }
                 else if(!tasks.contains(currentTask)) {
-                    updateDailyTask(key);
+                    dailyTasks.updateTask(key, selectTask(tasks), updateTimeMillis);
+                }
+                else if(updateTimeMillis != null && dailyTasks.shouldUpdate(key, updateTimeMillis)) {
+                    dailyTasks.updateTask(key, selectTask(tasks), updateTimeMillis);
                 }
             }
         }
-    }
-
-    // convenience method
-    private void updateDailyTask(String key) {
-        System.out.println("in1: " + key);
-        updateDailyTask(this.get(key));
-    }
-
-    // update a single task for today
-    private void updateDailyTask(ListOfTasks tasks) {
-        // choose a random task
-        String task = selectTask(tasks);
-
-        // store that task for today
-        pref.edit().putString(tasks.getFilename(), task).commit();
     }
 
     private String selectTask(ListOfTasks tasks) {
@@ -158,10 +163,61 @@ public class TasksManager extends TreeMap<String, ListOfTasks> implements Update
     }
     //#endregion
 
-    //#region low level file stuff
-    private void load() {
+    //#region handle type of storage
+    private void load(Context ...c) {
+        if(localMode) {
+            if(c == null || c.length == 0 || c[0] == null) {
+                throw new Error();
+            }
+
+            loadLocal(c[0]);
+        }
+        else {
+            loadDatabase();
+        }
+    }
+    //#endregion
+
+    //#region firebase control
+    private void loadDatabase() {
+        FirebaseDatabase database = FirebaseDatabase.getInstance();
+        dataRef = database.getReference("lists");
+        dataRef.addChildEventListener(new ChildEventListener() {
+            @Override
+            public void onChildAdded(@NonNull DataSnapshot dataSnapshot, @Nullable String s) {
+                ListOfTasks tasks = dataSnapshot.getValue(ListOfTasksFire.class);
+                TasksManager.this.put(dataSnapshot.getKey(), tasks);
+            }
+
+            @Override
+            public void onChildChanged(@NonNull DataSnapshot dataSnapshot, @Nullable String s) {
+
+            }
+
+            @Override
+            public void onChildRemoved(@NonNull DataSnapshot dataSnapshot) {
+
+            }
+
+            @Override
+            public void onChildMoved(@NonNull DataSnapshot dataSnapshot, @Nullable String s) {
+
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError databaseError) {
+
+            }
+        });
+    }
+    //#endregion
+
+    //#region local control
+    private void loadLocal(Context c) {
+        pref = c.getSharedPreferences("tasks",Context.MODE_PRIVATE);
+        setDirectory(c);
         for(String filename : fileNames()) {
-            ListOfTasks tasks = new ListOfTasks(getDirectory(), filename);
+            ListOfTasksLocal tasks = new ListOfTasksLocal(getDirectory(), filename);
             tasks.loadFromFile();
             this.put(filename, tasks);
             order.add(filename);
@@ -171,14 +227,12 @@ public class TasksManager extends TreeMap<String, ListOfTasks> implements Update
     private String[] fileNames() {
         return getDirectory().list();
     }
-
-    private File getDirectory() {
-        File directory = new File(context.getFilesDir(), "tasks");
+    private File getDirectory() { return directory; }
+    private void setDirectory(Context c) {
+        directory = new File(c.getFilesDir(), "tasks");
 
         // creates the directory if not present yet
         directory.mkdir();
-
-        return directory;
     }
     //#endregion
